requirements.txt
pymupdf==1.22.1
numpy==1.24.2
opencv-python==4.7.0.72
pydantic==1.10.7
pandas==2.0.1
flask==2.3.2
easyocr==1.7.0
torch==2.0.1
gunicorn==20.1.0
openai
python-dotenv==1.0.0
